fix gym.spaces.Box warning

+++ leap_c/examples/chain/env.py
@@ -66,15 +66,13 @@ class ChainEnv(gym.Env):
         vel_max = np.array([2.0, 2.0, 2.0] * (n_mass - 2))
         vel_min = -vel_max
         self.observation_space = spaces.Box(
-            low=np.concatenate([pos_min, vel_min]),
-            high=np.concatenate([pos_max, vel_max]),
-            dtype=np.float32,
+            low=np.concatenate([pos_min, vel_min], dtype=np.float32),
+            high=np.concatenate([pos_max, vel_max], dtype=np.float32),
         )
 
         self.action_space = spaces.Box(
-            low=np.array([-vmax, -vmax, -vmax]),
-            high=np.array([vmax, vmax, vmax]),
-            dtype=np.float32,
+            low=np.array([-vmax, -vmax, -vmax], dtype=np.float32),
+            high=np.array([vmax, vmax, vmax], dtype=np.float32),
         )
 
         self.render_mode = render_mode
